Test ServiceProvider against different EventDispatcher classes
