+++ tests/Laravel/Dispatcher.php
@@ -3,6 +3,6 @@ declare(strict_types=1);
 
 namespace Illuminate\Events;
 
-class Dispatcher extends EventDispatcher
+class Dispatcher extends \Tests\Laravel\EventDispatcherBase
 {
 }
