+++ tests/Laravel/EventDispatcher.php
@@ -3,28 +3,6 @@ declare(strict_types=1);
 
 namespace Illuminate\Events;
 
-class EventDispatcher
+class EventDispatcher extends \Tests\Laravel\EventDispatcherBase
 {
-    /**
-     * @var array
-     */
-    protected $listeners;
-
-    public function listen($name, $listener)
-    {
-        if (!isset($this->listeners[$name])) {
-            $this->listeners[$name] = [];
-        }
-
-        $this->listeners[$name][] = $listener;
-    }
-
-    public function dispatch($name, $event = null)
-    {
-        if (isset($this->listeners[$name])) {
-            foreach ($this->listeners[$name] as $listener) {
-                call_user_func($listener, $event);
-            }
-        }
-    }
 }
